Changes user and auth routes, and restructures user, auth, module, and problem routes test files with bcrypt hashing promist
