+++ __tests__/problem.routes.test.ts
@@ -21,23 +21,27 @@ describe('GET /', () => {
       });
     });
 
-    bcrypt.compare(pass, hashedPassword, async function (_err, result) {
-      try {
-        if (result) {
-          //User creation for token
-          await UserModel.create({
-            name: 'Test TA',
-            username: 'dhruv2000patel@gmail.com',
-            password: hashedPassword,
-            role: 'TA'
-          });
-        } else {
-          throw { message: 'Hash method not working properly' };
+    const result: Boolean = await new Promise((resolve) => {
+      bcrypt.compare(pass, hashedPassword, async function (_err, result) {
+        try {
+          return resolve(result);
+        } catch {
+          return false;
         }
-      } catch (err) {
-        return err;
-      }
+      });
     });
+
+    if (result) {
+      //User creation for token
+      await UserModel.create({
+        name: 'Test TA',
+        username: 'dhruv2000patel@gmail.com',
+        password: hashedPassword,
+        role: 'TA'
+      });
+    } else {
+      throw { message: 'Hash method not working properly' };
+    }
 
     const module = await Module.create({
       name: 'Stacks/Lists/Queues',
