+++ src/api/routes/v1/index.ts
@@ -1,6 +1,7 @@
 import healthRouter from './health.routes';
 import codeRouter from './code.routes';
 import authRouter from './auth.routes';
+import userRouter from './user.routes';
 import { studentProblemRouter, adminProblemRouter } from './problem.routes';
 import { moduleRouter } from './module.routes';
 import { submissionRouter } from './submission.routes';
@@ -13,7 +14,11 @@ router.use('/module', moduleRouter);
 router.use('/code/run', codeRouter);
 router.use('/student/problem', studentProblemRouter);
 router.use('/admin/problem', adminProblemRouter);
-router.use('/user', authRouter);
+//This is the old ones
+// router.use('/user', authRouter);
+//This is the new addition
+router.use('/user', userRouter);
+router.use('/auth', authRouter);
 router.use('/submission', submissionRouter);
 
 export default router;
